+++ npm-api-maker/src/collection.js
@@ -11,10 +11,14 @@ import uniqunize from "uniqunize"
  * @typedef {InstanceType<MC>} ModelOf
  */
 
+/** @typedef {Record<string, any>} PreloadMap */
+
+/** @typedef {string | Array<string | PreloadMap> | PreloadMap} PreloadValue */
+
 /**
  * @template {typeof import("./base-model.js").default} MC
  * @typedef {object} CollectionArgsType
- * @property {ModelOf<MC>} [model]
+ * @property {import("./base-model.js").default} [model]
  * @property {MC} modelClass
  * @property {string} [reflectionName]
  */
@@ -32,7 +36,7 @@ import uniqunize from "uniqunize"
  * @property {Record<string, any>} [params]
  * @property {number} [per]
  * @property {string} [perKey]
- * @property {string[]} [preload]
+ * @property {PreloadValue} [preload]
  * @property {Record<string, any>} [ransack]
  * @property {Record<string, any>} [search]
  * @property {string} [searchKey]
@@ -60,7 +64,7 @@ export default class ApiMakerCollection {
    * @returns {this}
    */
   abilities(originalAbilities) {
-    const newAbilities = /** @type {Record<string, string[]>} */ ({}) // eslint-disable-line no-extra-parens
+    const newAbilities = /** @type {Record<string, string[]>} */ ({})
 
     for (const originalAbilityName in originalAbilities) {
       const newModelName = inflection.underscore(originalAbilityName)
@@ -139,7 +143,7 @@ export default class ApiMakerCollection {
    * @returns {this}
    */
   groupBy(...arrayOfTablesAndColumns) {
-    const arrayOfTablesAndColumnsWithLowercaseColumns = /** @type {Array<string | [string, string]>} */ ([]) // eslint-disable-line no-extra-parens
+    const arrayOfTablesAndColumnsWithLowercaseColumns = /** @type {Array<string | [string, string]>} */ ([])
 
     for (const tableAndColumn of arrayOfTablesAndColumns) {
       if (Array.isArray(tableAndColumn)) {
@@ -277,7 +281,7 @@ export default class ApiMakerCollection {
   map(callback) { return this.loadedArray().map(callback) }
 
   /**
-   * @param {string[]} preloadValue
+   * @param {PreloadValue} preloadValue
    * @returns {this}
    */
   preload(preloadValue) {
@@ -376,7 +380,7 @@ export default class ApiMakerCollection {
 
     this._addQueryToModels(models)
 
-    const result = new Result(/** @type {any} */ ({collection: this, models, response})) // eslint-disable-line no-extra-parens
+    const result = new Result(/** @type {any} */ ({collection: this, models, response}))
 
     return result
   }
@@ -403,7 +407,7 @@ export default class ApiMakerCollection {
    * @returns {this}
    */
   select(originalSelect) {
-    const newSelect = /** @type {Record<string, string[]>} */ ({}) // eslint-disable-line no-extra-parens
+    const newSelect = /** @type {Record<string, string[]>} */ ({})
 
     for (const originalModelName in originalSelect) {
       const newModelName = inflection.underscore(originalModelName)
@@ -426,7 +430,7 @@ export default class ApiMakerCollection {
    * @returns {this}
    */
   selectColumns(originalSelect) {
-    const newSelect = /** @type {Record<string, string[]>} */ ({}) // eslint-disable-line no-extra-parens
+    const newSelect = /** @type {Record<string, string[]>} */ ({})
 
     for (const originalModelName in originalSelect) {
       const newModelName = inflection.underscore(inflection.underscore(originalModelName))