+++ npm-api-maker/src/table/filters/filter-form.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable comma-dangle, implicit-arrow-linebreak, import/no-named-as-default */
-/* eslint-disable complexity, new-cap, newline-per-chained-call, no-await-in-loop, no-extra-parens, no-return-assign */
+/* eslint-disable complexity, new-cap, newline-per-chained-call, no-await-in-loop, no-return-assign */
 /* eslint-disable no-unused-vars, react/jsx-max-depth, react/jsx-no-literals */
 /* eslint-disable react/jsx-sort-props, react/no-array-index-key, sort-imports */
 import React, {useMemo, useRef} from "react"