+++ npm-api-maker/src/use-input.js
@@ -1,11 +1,11 @@
-/* eslint-disable implicit-arrow-linebreak, no-extra-parens, no-use-before-define, prefer-object-spread */
+/* eslint-disable implicit-arrow-linebreak, no-use-before-define, prefer-object-spread */
 import {dig, digg} from "diggerize"
 import {useCallback, useEffect, useMemo, useRef} from "react"
 import idForComponent from "./inputs/id-for-component.js"
 import nameForComponent from "./inputs/name-for-component.js"
 import strftime from "strftime"
 import {useForm} from "./form"
-import useShape from "set-state-compare/build/use-shape.js"
+import useShape from "./use-shape.js"
 import useValidationErrors from "./use-validation-errors.js"
 
 /**